farm_system/requirements.txt
django
django-polymorphic
asgiref